refactor(site): simplify the homepage hero and use ut installation
