+++ docs/index.md
@@ -5,7 +5,7 @@ toc: false
 hero:
   title: Fast linting.
   accent: Familiar rules.
-  description: Find problems in JavaScript and TypeScript with a native Zig engine. Keep the rule names you know, configure in TypeScript, and apply safe fixes from your terminal to CI.
+  description: A fast JavaScript and TypeScript linter, built in Zig.
 ---
 
 <ProductHome />
